pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-triage-env"
version = "0.2.0"
description = "Personal Financial Triage & Life Budget Advisor — an OpenEnv environment for multi-day financial decision-making."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.109.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.3.0",
]

[tool.setuptools.packages.find]
include = ["server*"]
